refactor: 2단계 요구사항 충족
